weird bug that may cause an transaction exception

+++ redis/redis-console/src/main/java/com/ctrip/xpipe/redis/console/dao/ClusterDao.java
@@ -93,11 +93,11 @@ public class ClusterDao extends AbstractXpipeConsoleDAO{
 			}
 		});
 		
-		if(null != shards) {
+		if(null != shards && !shards.isEmpty()) {
 			shardDao.deleteShardsBatch(shards);
 		}
 		
-		if(null != dcClusters) {
+		if(null != dcClusters && !dcClusters.isEmpty()) {
 			dcClusterDao.deleteDcClustersBatch(dcClusters);
 		}
